Potentially working PoC for using volumentric texture to store light data for voxels (need ogniK to rescue me on the shader side)

+++ code/blocks/GlassBlock.cs
@@ -10,6 +10,6 @@ namespace Facepunch.CoreWars.Blocks
 		public override bool IsTranslucent => false;
 		public override byte TextureId => 1;
 		public override byte BlockId => 4;
-		public override int LightLevel => 14;
+		public override byte LightLevel => 14;
 	}
 }
